requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
flask-openapi3==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
